Move validation exception to canonical API package
