+++ klum-ast-bean-validation/src/test/groovy/com/blackbuild/klum/ast/validation/bean/TransformedBeanValidationTest.groovy
@@ -24,7 +24,7 @@
 package com.blackbuild.klum.ast.validation.bean
 
 import com.blackbuild.klum.ast.AbstractDSLSpec
-import com.blackbuild.klum.ast.runtime.KlumValidationException
+import com.blackbuild.klum.ast.runtime.validation.KlumValidationException
 import spock.lang.Issue
 
 @Issue("391")
